% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ap_wp.R
\name{ap_wp}
\alias{ap_wp}
\title{Find Articulation points and weak point of a graph}
\usage{
ap_wp(g, type = "weak")
}
\arguments{
\item{g}{a connected igraph graph}

\item{type}{connectedness type for directed graphs : "weak" or "strong". Ignored for undirected graphs}
}
\value{
A list with 3 elements :
\itemize{
\item \code{ap} is a character vector of articulation points names
\item \code{ap_1} is a character vector of 1-order articulation points names
\item \code{wp} is a character vector of weak points names
}
}
\description{
Find the articulation points (AP), the 1-order articulation
points (1-AP) and weak points of a graph.
}
\details{
Articulation points are determined with the \code{\link[igraph]{articulation_points}} function.

A 1-order articulation point is an articulation point which generates a component of cardinal 1
when removed.

A weak point is a singleton component of a graph when one of its 1-order articulation points is
removed.
}
\references{
Monique Dalut-Vincent, Michel Forsé, Jean-Paul Auray, "An algorithm for finding the structure of social groups", \emph{Social Networks}, 16 (1994) 137-162.
}

