% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reso-native.R
\name{s}
\alias{s}
\alias{p}
\alias{d}
\alias{x}
\title{reso-native functions}
\usage{
s(string, i = 1)

p(...)

d(...)

x(string)
}
\arguments{
\item{string}{A character value}

\item{i}{An integer value}

\item{...}{Passed to \code{paste0} for \code{p()} or \code{deparse()} for \code{d()}.}
}
\description{
This package contains four wrapper functions that make almost all of R
available, by allowing basic string manipulation and evaluation.
}
\section{Functions}{
\itemize{
\item \code{s}: Given a \code{string} and a number \code{i}, returns the character at position \code{i} from \code{string}

\item \code{p}: An alias for \code{paste0(...,collapse = "")}; essentially squashes
items together into a single string.

\item \code{d}: An alias for \code{deparse(...)}; converts an R expression into a literal string.

\item \code{x}: Evaluates a given string as an R expression.
}}

\examples{
s(c()) # character(0)
d(character(0)) # "character(0)"
c(p()) # ""
x("1 + 1") # 2
}
\keyword{functions}
\keyword{native}
