% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{plot_reso}
\alias{plot_reso}
\title{Plot a graph with vertices colored by a group variable}
\usage{
plot_reso(g, group)
}
\arguments{
\item{g}{an igraph graph}

\item{group}{vector of vertices group values}
}
\description{
Plot a graph with vertices colored by a group variable
}
\examples{
## Random graph
g <- igraph::sample_gnp(14, 0.35)
plot(g)
res <- reso(g)
plot_reso(g, res)

}

