% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reso.R
\name{reso}
\alias{reso}
\title{Apply the RESO algorithm to a graph}
\usage{
reso(g)
}
\arguments{
\item{g}{an igraph graph}
}
\value{
Character vector with each vertex group. Groups are given in the form XX_YY,
where XX is the type of group ("WP" for weak points, "1-AP" for 1-order
articulation points, "CC" for connected components, and "CCsing" for
singleton connected components), and YY the level of the group (the iteration
during which the group has been produced).
}
\description{
Apply the RESO algorithm to a graph
}
\details{
See the reference article for details and explanations about the algorithm.
}
\examples{
## Random graph
g <- igraph::sample_gnp(26, 0.15)
plot(g)
res <- reso(g)
plot_reso(g, res)
}
\references{
Monique Dalut-Vincent, Michel Forsé, Jean-Paul Auray, "An algorithm for finding the structure of social groups", \emph{Social Networks}, 16 (1994) 137-162.
}

