% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reso.R
\name{reso_decompose}
\alias{reso_decompose}
\title{RESO algorithm decomposition iteration}
\usage{
reso_decompose(g, iter, ap1, ap, type)
}
\arguments{
\item{g}{an igraph graph}

\item{iter}{iteration number}

\item{ap1}{list of previous non clustered 1-AP}

\item{ap}{list of previous non clustered ap}
}
\description{
Not to be called directly
}

