% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reso_source.R
\name{reso_source}
\alias{reso_source}
\title{reso_source()}
\usage{
reso_source(file)
}
\arguments{
\item{file}{A file or a string, to be executed as a reso expression.}
}
\value{
The results of the evaluated reso expression.
}
\description{
Execute the contents of a file or string as a reso
expression. This just validates the file's contents
with reso_validate() and then interprets the string
as a plain R expression with \code{eval(parse(text=file))}.
}
\examples{
reso_source("x(s(d(p()[0])))")
reso_source("x(s(d(p()[0])))(0,x(s(d(p()[0])))(!0,0[0]))")
filename <- system.file("extdata", "example.reso", package = "reso")
reso_source(filename)
}
