% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reso-native.R
\name{s}
\alias{s}
\title{reso-native functions}
\usage{
s(string, i = 1)
}
\description{
This package contains four wrapper functions that make almost all of R
available. They allow basic string manipulation and evaluation.
}
\examples{
s(c()) # character(0)
c(p()) # ""
d(character(0)) # "character(0)"
x("1 + 1") # 2
}
\keyword{functions}
